testing db

+++ web_server/src/app.py
@@ -120,7 +120,7 @@ def detect_faces_in_image(file_stream):
         # See if the first face in the uploaded image matches the known face of Obama
 
         for row in Row.select():
-            match_results = face_recognition.compare_faces(np.fromstring(row.img_encoding), unknown_face_encodings[0])
+            match_results = face_recognition.compare_faces(np.fromstring(row.img_encoding, dtype=np.uint8), unknown_face_encodings[0])
             if match_results[0]:
                 result = row.user_id
                 break
